cut map page, move map to component, add paint styles
